sqle/indexmeta: extract LookupMeta into a leaf package so dsess no longer imports sqle/index
SessionCache held []index.LookupMeta, which forced dsess to import
sqle/index and pull in the entire go-mysql-server query planner via
its transitive deps. LookupMeta is a small four-field value type with
no behavior, so it moves to a new sqle/indexmeta package that depends
only on go-mysql-server/sql. sqle/index keeps a type alias for
backwards compatibility with existing callers.
