+++ go/libraries/doltcore/sqle/dsess/session_cache.go
@@ -21,7 +21,7 @@ import (
 	"github.com/dolthub/go-mysql-server/sql"
 
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
-	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/index"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/indexmeta"
 )
 
 // CacheableDoltTable is implemented by tables in SessionCache.tables.
@@ -52,7 +52,7 @@ type SessionCache struct {
 	// writers are keyed by table schema hash
 	writers map[doltdb.DataCacheKey]*WriterState
 	// strictLookups are keyed by table schema hash
-	strictLookups map[doltdb.DataCacheKey][]index.LookupMeta
+	strictLookups map[doltdb.DataCacheKey][]indexmeta.LookupMeta
 	// checks is keyed by table schema hash
 	checks map[doltdb.DataCacheKey][]sql.CheckDefinition
 
@@ -202,7 +202,7 @@ func (c *SessionCache) GetCachedTable(key doltdb.DataCacheKey, tableName TableCa
 }
 
 // GetCachedStrictLookup returns a cached set of |sql.IndexLookup|s for the table named, and whether the cache was present
-func (c *SessionCache) GetCachedStrictLookup(key doltdb.DataCacheKey) ([]index.LookupMeta, bool) {
+func (c *SessionCache) GetCachedStrictLookup(key doltdb.DataCacheKey) ([]indexmeta.LookupMeta, bool) {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	schemaState, ok := c.strictLookups[key]
@@ -210,12 +210,12 @@ func (c *SessionCache) GetCachedStrictLookup(key doltdb.DataCacheKey) ([]index.L
 }
 
 // CacheStrictLookup caches a set of |sql.IndexLookup|s for the table named
-func (c *SessionCache) CacheStrictLookup(key doltdb.DataCacheKey, state []index.LookupMeta) {
+func (c *SessionCache) CacheStrictLookup(key doltdb.DataCacheKey, state []indexmeta.LookupMeta) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 
 	if c.strictLookups == nil {
-		c.strictLookups = make(map[doltdb.DataCacheKey][]index.LookupMeta)
+		c.strictLookups = make(map[doltdb.DataCacheKey][]indexmeta.LookupMeta)
 	}
 	if len(c.strictLookups) > maxCachedKeys {
 		for k := range c.strictLookups {
