+++ go/libraries/doltcore/sqle/index/dolt_index.go
@@ -31,6 +31,7 @@ import (
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb/durable"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema/typeinfo"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/indexmeta"
 	"github.com/dolthub/dolt/go/store/datas"
 	"github.com/dolthub/dolt/go/store/hash"
 	"github.com/dolthub/dolt/go/store/pool"
@@ -564,12 +565,8 @@ type doltIndex struct {
 	fulltext bool
 }
 
-type LookupMeta struct {
-	Idx      sql.Index
-	Fds      *sql.FuncDepSet
-	Cols     sql.FastIntSet
-	Ordinals []int
-}
+// LookupMeta is an alias for indexmeta.LookupMeta kept here for callers within sqle.
+type LookupMeta = indexmeta.LookupMeta
 
 // GetStrictLookups searches |indexes| and returns a slice of LookupMeta that indicates which indexes can be used
 // for strict lookups. Indexes may be used for strict lookups if they are a unique index and the columns they
